+++ internal/provider/provider.go
@@ -61,8 +61,8 @@ func (p *cpanelProvider) Metadata(_ context.Context, _ provider.MetadataRequest,
 // Schema defines the provider-level schema for configuration data.
 func (p *cpanelProvider) Schema(_ context.Context, _ provider.SchemaRequest, resp *provider.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		Description:         "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, ModSecurity settings, HTTP redirects, directory indexes and privacy, Git repositories, custom MIME types and Apache handlers, email accounts, forwarders, autoresponders, FTP accounts, website IP blocks, remote database hosts, and MySQL, MariaDB, and PostgreSQL users and databases.",
-		MarkdownDescription: "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, ModSecurity settings, HTTP redirects, directory indexes and privacy, Git repositories, custom MIME types and Apache handlers, email accounts, forwarders, autoresponders, FTP accounts, website IP blocks, remote database hosts, and MySQL, MariaDB, and PostgreSQL users and databases.",
+		Description:         "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, ModSecurity settings, HTTP redirects, directory indexes and privacy, Git repositories, custom MIME types and Apache handlers, email accounts and suspensions, forwarders, autoresponders, FTP accounts, website IP blocks, remote database hosts, and MySQL, MariaDB, and PostgreSQL users and databases.",
+		MarkdownDescription: "Manage cPanel account API tokens, cron jobs, DNS records, Dynamic DNS domains, web domains, ModSecurity settings, HTTP redirects, directory indexes and privacy, Git repositories, custom MIME types and Apache handlers, email accounts and suspensions, forwarders, autoresponders, FTP accounts, website IP blocks, remote database hosts, and MySQL, MariaDB, and PostgreSQL users and databases.",
 		Attributes: map[string]schema.Attribute{
 			"username": schema.StringAttribute{
 				Optional:            true,
@@ -276,6 +276,7 @@ func (p *cpanelProvider) DataSources(_ context.Context) []func() datasource.Data
 		NewDomainAliasDataSource,
 		NewDynamicDNSDataSource,
 		NewEmailAccountDataSource,
+		NewEmailAccountSuspensionDataSource,
 		NewEmailAutoResponderDataSource,
 		NewEmailDomainForwarderDataSource,
 		NewEmailForwarderDataSource,
@@ -308,6 +309,7 @@ func (p *cpanelProvider) Resources(_ context.Context) []func() resource.Resource
 		NewDomainAliasResource,
 		NewDynamicDNSResource,
 		NewEmailAccountResource,
+		NewEmailAccountSuspensionResource,
 		NewEmailAutoResponderResource,
 		NewEmailDomainForwarderResource,
 		NewEmailForwarderResource,